refactor(admin): convert sign in component to new naming convention, add sign in errors as a hook
