+++ apps/admin.enes.sh/app/features/auth/index.ts
@@ -1,1 +1,1 @@
-export * from './SignIn'
+export * from './sign-in'
