+++ apps/admin.enes.sh/app/features/auth/sign-in-form.tsx
@@ -1,0 +1,16 @@
+import { Button } from '@enes-sh/ui'
+
+export type SignInFormTypes = {
+  disabled: boolean
+  onSubmit: () => Promise<void>
+}
+
+const SignInForm = ({ disabled = false, onSubmit }: SignInFormTypes) => {
+  return (
+    <Button className='w-full text-sm' onClick={onSubmit} disabled={disabled}>
+      Sign In
+    </Button>
+  )
+}
+
+export { SignInForm }
