+++ apps/admin.enes.sh/app/hooks/use-sign-in-errors.tsx
@@ -1,0 +1,18 @@
+const useSignInErrors = (): Record<string, string> => {
+  return {
+    AccessDenied: 'Access denied. Please retry again.',
+    Signin: 'Try signing in with a different account.',
+    OAuthSignin: 'Try signing in with a different account.',
+    OAuthCallback: 'Try signing in with a different account.',
+    OAuthCreateAccount: 'Try signing in with a different account.',
+    EmailCreateAccount: 'Try signing in with a different account.',
+    Callback: 'Try signing in with a different account.',
+    OAuthAccountNotLinked:
+      'To confirm your identity, sign in with the same account you used originally.',
+    EmailSignin: 'The e-mail could not be sent.',
+    CredentialsSignin: 'Sign in failed. Check the details you provided are correct.',
+    SessionRequired: 'Please sign in to access this page.',
+    default: 'Unable to sign in.'
+  }
+}
+export { useSignInErrors }
